CODE RUB: organize retrieve profile id tests.
